Cambiar dias habiles/working_days de long a double y agregar converter
- SolicitudDTO.diasHabiles y working_days ahora son double
- Nuevo FlexibleNullableDoubleConverter
- Comentar campos no usados en VacacionesRest
- Agregar logging en ColaboradorService y AsistenciaRepository

+++ DTOs/SolicitudDTO.cs
@@ -15,7 +15,7 @@ public class SolicitudDTO
   public DateTime fechaAutorizacion { get; set; }
   public long id_autorizo { get; set; }
 
-  public long diasHabiles { get; set; }
+  public double diasHabiles { get; set; }
   
 }
